Move hooks and contexts into lib
