+++ src/lib/hooks/use-low-power-mode.ts
